sqlite integration done

+++ RentVillaAPI/Core/RentVilla.Application/DTOs/CartDTOs/AddCartItemDTO.cs
@@ -15,7 +15,7 @@ namespace RentVilla.Application.DTOs.CartDTOs
         public int AdultNumber { get; set; }
         public int ChildrenNumber { get; set; }
         public string Note { get; set; } = null;
-        public decimal Price { get; set; }
-        public decimal TotalCost { get; set; }
+        public double Price { get; set; }
+        public double TotalCost { get; set; }
     }
 }
