+++ RentVillaAPI/Core/RentVilla.Application/DTOs/ProductDTOs/UpdateProductDTO.cs
@@ -12,8 +12,8 @@ namespace RentVilla.Application.DTOs.ProductDTOs
         public string Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
-        public decimal Deposit { get; set; }
+        public double Price { get; set; }
+        public double Deposit { get; set; }
         public string MapId { get; set; }
         public string Address { get; set; }
         public ProductAddressDTO ProductAddress { get; set; }
