+++ RentVillaAPI/Core/RentVilla.Application/DTOs/ReservationDTOs/GetReservationDTO.cs
@@ -18,8 +18,8 @@ namespace RentVilla.Application.DTOs.ReservationDTOs
         public int AdultNumber { get; set; }
         public int ChildrenNumber { get; set; }
         public string Note { get; set; }
-        public decimal ProductPrice { get; set; }
-        public decimal TotalCost { get; set; }
+        public double ProductPrice { get; set; }
+        public double TotalCost { get; set; }
         public bool IsPaid { get; set; }
         public string ConversationId { get; set; }
         public string PaymentId { get; set; }
