+++ RentVillaAPI/Core/RentVilla.Application/Feature/Commands/Products/CreateProduct/CreateProductCommandRequest.cs
@@ -13,8 +13,8 @@ namespace RentVilla.Application.Feature.Commands.Products.CreateProduct
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
-        public decimal Deposit { get; set; }
+        public double Price { get; set; }
+        public double Deposit { get; set; }
         public string MapId { get; set; }
         public string Address { get; set; }
         public ProductAddressDTO ProductAddress { get; set; }
