+++ RentVillaAPI/Core/RentVilla.Application/ViewModels/Reservation/AddReservationVM.cs
@@ -14,7 +14,7 @@ namespace RentVilla.Application.ViewModels.Reservation
         public int AdultNumber { get; set; }
         public int ChildrenNumber { get; set; }
         public string Note { get; set; }
-        public decimal ProductPrice { get; set; }
-        public decimal TotalCost { get; set; }
+        public double ProductPrice { get; set; }
+        public double TotalCost { get; set; }
     }
 }
